requirements.txt
EasyTkinter==1.1.0
Pillow==8.0.1
opencv-python==4.4.0.44